# docs/swagger.yaml
basePath: /
definitions:
  handler.DashboardStats:
    properties:
      active_trucks:
        type: integer
      order_breakdown:
        properties:
          cancelled:
            type: integer
          completed:
            type: integer
          partial:
            type: integer
          pending:
            type: integer
        type: object
      total_admins:
        type: integer
      total_orders:
        type: integer
      total_trucks:
        type: integer
      total_users:
        type: integer
    type: object
...
  model.CreateDriverRequest:
    properties:
      is_active:
        example: true
        type: boolean
      license_number:
        example: SIM-B-1234567
        type: string
      name:
        example: Andi Wijaya
        type: string
      phone:
        example: "+6281122334455"
        type: string
      status:
        example: available
        type: string
    required:
    - license_number
    - name
    - phone
    type: object
...
  model.UpdateDriverRequest:
    properties:
      is_active:
        example: true
        type: boolean
      license_number:
        example: SIM-B-1234567
        type: string
      name:
        example: Andi Wijaya
        type: string
      phone:
        example: "+6281122334455"
        type: string
      status:
        example: available
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@jalurberlian.id
    name: API Support
    url: http://www.jalurberlian.id/support
  description: Real-time fleet management and order tracking system for PT. Jalur
    Berlian Makassar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Jalur Berlian Fleet Management API
  version: "1.0"
paths:
...
  /admin/dashboard/stats:
    get:
      consumes:
      - application/json
      description: Get dashboard overview with order stats, truck count, user count,
        etc.
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard statistics
          schema:
            $ref: '#/definitions/handler.DashboardStats'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get dashboard statistics
      tags:
      - Admin
  /admin/drivers:
    get:
      description: Get list of all drivers. All authenticated admins can access.
      produces:
      - application/json
      responses:
        "200":
          description: Driver list with count
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List drivers
      tags:
      - Drivers
    post:
      consumes:
      - application/json
      description: Create a new driver. All authenticated admins can access.
      parameters:
      - description: 'Driver data: name, license_number, phone, status, is_active'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateDriverRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Driver created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create driver
      tags:
      - Drivers
  /admin/drivers/{id}:
    delete:
      description: Deactivate a driver by ID. All authenticated admins can access.
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Driver deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - invalid driver ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found - driver tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete driver
      tags:
      - Drivers
    get:
      description: Retrieve a driver by ID.
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Driver detail
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid driver ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found - driver tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get driver detail
      tags:
      - Drivers
    patch:
      consumes:
      - application/json
      description: Update driver fields (partial). All authenticated admins can access.
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: integer
      - description: Driver update data (partial)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateDriverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Driver updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found - driver tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update driver
      tags:
      - Drivers
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"