# internal/model/driver.go
package model

// Driver represents truck driver data.
type Driver struct {
	ID            int    `db:"id" json:"id"` // Immutable: Auto-generated by database
	Name          string `db:"name" json:"name" example:"Andi Wijaya"`
	LicenseNumber string `db:"license_number" json:"license_number" example:"SIM-B-1234567"`
	Phone         string `db:"phone" json:"phone" example:"+6281122334455"`
	Status        string `db:"status" json:"status" example:"available"`
	IsActive      bool   `db:"is_active" json:"is_active" example:"true"`
}

// CreateDriverRequest is the request DTO for creating a new driver.
type CreateDriverRequest struct {
	Name          string `json:"name" example:"Andi Wijaya" binding:"required"`
	LicenseNumber string `json:"license_number" example:"SIM-B-1234567" binding:"required"`
	Phone         string `json:"phone" example:"+6281122334455" binding:"required"`
	Status        string `json:"status" example:"available"`
	IsActive      bool   `json:"is_active" example:"true"`
}

// UpdateDriverRequest is the request DTO for updating a driver (partial update).
type UpdateDriverRequest struct {
	Name          *string `json:"name" example:"Andi Wijaya"`
	LicenseNumber *string `json:"license_number" example:"SIM-B-1234567"`
	Phone         *string `json:"phone" example:"+6281122334455"`
	Status        *string `json:"status" example:"available"`
	IsActive      *bool   `json:"is_active" example:"true"`
}